+++ src/experiments/train_models_competing_risks.py
@@ -44,7 +44,7 @@ torch.cuda.manual_seed_all(0)
 random.seed(0)
 
 # Set precision
-dtype = torch.float64
+dtype = torch.float32
 torch.set_default_dtype(dtype)
 
 # Setup device
@@ -167,8 +167,8 @@ if __name__ == "__main__":
             model = util.get_model_and_output("hierarch_full", train_data, test_data,
                                               valid_data, config, hyperparams, verbose)
         elif model_name == "mtlrcr":
-            train_events = train_dict['E'].type(torch.int64).cpu().numpy()
-            valid_events = valid_dict['E'].type(torch.int64).cpu().numpy()
+            train_events = train_dict['E'].type(torch.int32).cpu().numpy()
+            valid_events = valid_dict['E'].type(torch.int32).cpu().numpy()
             y_train = encode_mtlr_format(train_dict['T'], train_events, time_bins.cpu().numpy())
             y_valid = encode_mtlr_format(valid_dict['T'], valid_events, time_bins.cpu().numpy())            
             num_time_bins = len(time_bins.cpu().numpy()) + 1
